pkg/client: restructure folders
