+++ pkg/client/v1/receipt.go
@@ -1,4 +1,4 @@
-package client
+package v1
 
 import (
 	"context"
